pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = "Platforme de gestion des taches"
authors = ["SALHI Mohamed <salhi.mohamed.sd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
psycopg2 = "^2.9.10"
httpx = "^0.28.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
